+++ tests/testthat/test-pwrcontrast.r
@@ -154,7 +154,8 @@ test_that("pwrcontrast matches expected results", {
 
     testthat::expect_equal(
       as.numeric(res_n),
-      as.numeric(e[,c("n_total", "ncp", "df_num", "df_denom")])
+      as.numeric(e[,c("n_total", "ncp", "df_num", "df_denom")]),
+      tolerance = 10e-5
     )
 
     # Alpha